refactor: change settings route structure
