+++ src/routes/index.ts
@@ -1,4 +1,5 @@
-export * from './routes-data';
-export * from './AppRouter';
-export * from './AuthRoutes';
-export * from './PrivateRoutes';
+export * from "./private-routes-data";
+export * from "./auth-routes-data";
+export * from "./AppRouter";
+export * from "./AuthRoutes";
+export * from "./PrivateRoutes";
