+++ src/settings/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./InputSetting";
+export * from "./SettingForm";
