+++ src/shared/components/index.ts
@@ -3,14 +3,12 @@ export * from "./CommentCard";
 export * from "./CommentsModal";
 export * from "./FollowButton";
 export * from "./InputComponent";
-export * from "./InputSetting";
 export * from "./LeftSidebar";
 export * from "./LeftSidebarLinks";
 export * from "./MainContainer";
 export * from "./ModalContainForm";
 export * from "./NavLink";
 export * from "./RightSidebar";
-export * from "./SettingForm";
 export * from "./Sidebar";
 export * from "./UserCard";
 export * from "./FormControlContainer";
